makes rendering identical document idempotent
Creating a different file identifier is painful for testing. While it
makes sense that files created at diffent times with the same content
have different file identifiers, both pdfkit and pdfmake allow to
overwrite the CreationDate and the various other parts that are used to
derive the pdf id.

This commit ensures idempotent rendering for users who explicitely
overwrite the creation date. This allows for an easier end2end testing.